Pass minimum quality duration to adaptive track selection.

+++ core/src/main/java/com/novoda/noplayer/Options.java
@@ -3,9 +3,9 @@ package com.novoda.noplayer;
 public class Options {
 
     private final ContentType contentType;
-    private final long minDurationBeforeQualityIncreaseInMillis;
+    private final int minDurationBeforeQualityIncreaseInMillis;
 
-    Options(ContentType contentType, long minDurationBeforeQualityIncreaseInMillis) {
+    Options(ContentType contentType, int minDurationBeforeQualityIncreaseInMillis) {
         this.contentType = contentType;
         this.minDurationBeforeQualityIncreaseInMillis = minDurationBeforeQualityIncreaseInMillis;
     }
@@ -14,7 +14,7 @@ public class Options {
         return contentType;
     }
 
-    public long minDurationBeforeQualityIncreaseInMillis() {
+    public int minDurationBeforeQualityIncreaseInMillis() {
         return minDurationBeforeQualityIncreaseInMillis;
     }
 }
